Pass timestamp through VersionService.insertNewVersion
The previous interface did not specify where the timestamp should come
from. This left it open to the implementation to use, say, database
timestamps, but when the database is on another host, its clock need not
be synchronized with ours and DocumentFilesResource.handleUpdate is not
reliable.

Making the timestamp part of the VersionService interface makes it
explicit that we use timestamps from the machine running textrepo-app.