[TT-481] Add a task to DELETE a document
Also added some more cascading deletes to the postgres tables. The only thing that cannot be cascaded and might be left orphaned right now are the actual 'contents' as these could be shared by another version.

TODO: figure out a sensible 'result'. Perhaps a list of the sha224 that were referenced?