Set Content-Type to original mimetype where feasible
When asking for contents by sha, mimetype is still application/octet-stream as contents can be shared among files which don't necessarily agree on the mimetype.

When asking for contents upwards from a given Version context, we can find the file it belongs to and yield the accompanying mimetype of that file.