More nuanced handling of new vs. existing contents on import
200 OK when no new version created vs.
201 CREATED when new version was created.

* isNewVersion in result object also discloses this

New query param 'asLatestVersion' determines whether it is ok if the currently uploaded file already exists as /any/ version or whether it has to be the /latest/ version.

If any version is ok, no new version will be made if contents of current upload are already available as any version of this file.

if 'asLatestVersion' is requested, new version will be made of this contents, *unless* these contents are already the current latest version.